# Conflicts:
#	.gradle/8.12.1/checksums/checksums.lock
#	.gradle/8.12.1/checksums/md5-checksums.bin
#	.gradle/8.12.1/checksums/sha1-checksums.bin
#	.gradle/8.12.1/executionHistory/executionHistory.lock
#	.gradle/8.12.1/fileHashes/fileHashes.bin
#	.gradle/8.12.1/fileHashes/fileHashes.lock
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	build.gradle.kts
#	src/main/java/be/kdg/integration4/domain/BikeReport.java
#	src/main/java/be/kdg/integration4/repository/BikeReportRepository.java
#	src/main/java/be/kdg/integration4/service/BikeReportService.java
#	src/main/java/be/kdg/integration4/service/BikeReportServiceImpl.java
#	src/main/resources/application.properties

+++ src/main/java/be/kdg/integration4/domain/TestBench.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class TestBench {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer benchId;
+    private Long benchId;
 
     @OneToOne
     private Workshop workshop;
